+++ src/services/face_reid.py
@@ -3,12 +3,15 @@ import os
 import pickle
 import time
 from enum import Enum
-from typing import List
+from typing import List, Optional
 
 import httpx
 from loguru import logger
+from tinydb import Query, TinyDB
+from tinydb.middlewares import CachingMiddleware
+from tinydb.storages import JSONStorage
 
-from ..utils.model_pydantic import ImageMetadata
+from ..utils.model_pydantic import ImageFaceRecognitionStatus, ImageMetadata
 from .faces_db_service import Face, FaceDBService
 from .redis_service import (
     RedisInterface,  # Assuming this is the file where RedisInterface is defined
@@ -56,20 +59,22 @@ class FaceRecognitionService:
         redis_interface: RedisInterface,
         face_db_service: FaceDBService,
         progress_file="face_recognition_progress.pkl",
+        db_path="face_recognition_progress.json",
     ):
         """
         Initializes the FaceRecognitionService.
 
         Args:
             redis_interface (RedisInterface): An instance of RedisInterface for data storage.
-            progress_file (str): Path to the pickle file for progress persistence.
+            db_path (str): Path to the TinyDB file for progress persistence.
+            pickle_file (str): Path to the existing pickle file for migration.
         """
         self.redis_interface = redis_interface
         self._face_db_service = face_db_service
 
         self.status = ProcessStatus.IDLE
         self.progress = 0
-        self.images: List[ImageMetadata] = []
+        # self.images: List[ImageMetadata] = []
         self.processed_images_names = []
         self.failed_images_names = []
         self._progress_file = progress_file
@@ -80,24 +85,56 @@ class FaceRecognitionService:
         self._terminate = False
         self._processing_task = None
 
+        # Initialize TinyDB
+        self._db = TinyDB(db_path, storage=CachingMiddleware(JSONStorage))
+
     async def load_progress(self):
-        logger.info("Start loading recognition progress data")
-        if os.path.exists(self._progress_file):
-            try:
-                with open(self._progress_file, "rb") as file:
-                    progress_data = pickle.load(file)
-                    self.images = [
-                        ImageMetadata(**image)
-                        for image in progress_data.get("images", [])
-                    ]
-                    self.processed_images_names = progress_data.get(
-                        "processed_images", []
-                    )
-                    self.progress = progress_data.get("progress", 0)
-                    self.failed_images_names = progress_data.get("failed_images", [])
-                    logger.info("Progress loaded. Resuming from the last saved state.")
-            except (pickle.PickleError, EOFError) as e:
-                print(f"Failed to load progress file: {e}")
+        """
+        Loads progress data from TinyDB.
+        """
+        logger.info("Loading progress from TinyDB.")
+
+        # Load metadata
+        progress_query = Query()
+        progress_data = self._db.get(progress_query.id == "progress_metadata")
+        if progress_data:
+            self.processed_images_names = progress_data.get("processed_images", [])
+            self.progress = progress_data.get("progress", 0)
+            self.failed_images_names = progress_data.get("failed_images", [])
+
+        logger.info("Progress loaded successfully.")
+
+    def get_images_count(self):
+        return self._db.count(Query().full_client_path.exists())
+
+    def migrate_pickle_to_tinydb(self):
+        """
+        Migrates data from the existing pickle file to the TinyDB database.
+        """
+        if not os.path.exists(self._progress_file):
+            logger.info("No pickle file found. Migration skipped.")
+            return
+
+        logger.info("Starting migration from pickle to TinyDB.")
+        try:
+            with open(self._progress_file, "rb") as file:
+                progress_data = pickle.load(file)
+
+            # Load data into the service attributes
+            images = [
+                ImageMetadata(**image) for image in progress_data.get("images", [])
+            ]
+            self.processed_images_names = progress_data.get("processed_images", [])
+            self.progress = progress_data.get("progress", 0)
+            self.failed_images_names = progress_data.get("failed_images", [])
+
+            # Use persist_progress to save data to TinyDB
+            self.load_images(images)
+            self.persist_progress()
+
+            logger.info("Migration completed successfully.")
+        except (pickle.PickleError, EOFError, FileNotFoundError) as e:
+            logger.error(f"Failed to migrate data from pickle: {e}")
 
     def load_images(self, images: List[ImageMetadata]) -> int:
         """
@@ -106,10 +143,18 @@ class FaceRecognitionService:
         Args:
             images (list): List of image paths to be processed.
         """
-        new_images = [image for image in images if image not in self.images]
-        self.images.extend(new_images)
-        self.persist_progress()
-        return len(new_images)
+        loaded_images = 0
+        for image in images:
+            if image.full_client_path in self.processed_images_names:
+                image.face_recognition_status = ImageFaceRecognitionStatus.DONE
+            if image.full_client_path in self.failed_images_names:
+                image.face_recognition_status = ImageFaceRecognitionStatus.FAILED
+            self._db.upsert(
+                image.model_dump(), Query().full_client_path == image.full_client_path
+            )
+            loaded_images += 1
+        self._db.storage.flush()
+        return loaded_images
 
     def get_status(self) -> dict:
         """
@@ -119,7 +164,7 @@ class FaceRecognitionService:
             tuple: (Status, progress percentage)
         """
         images_in_queue = (
-            len(self.images)
+            self.get_images_count()
             - len(self.processed_images_names)
             - len(self.failed_images_names)
         )
@@ -132,8 +177,10 @@ class FaceRecognitionService:
         return {
             "status": self.status,
             "progress": self.progress,
-            "images": len(self.images),
             "time_left_seconds": time_left,
+            "images": self.get_images_count(),
+            "processed_images": len(self.processed_images_names),
+            "failed_images": len(self.failed_images_names),
         }
 
     def stop(self):
@@ -158,90 +205,170 @@ class FaceRecognitionService:
             raise RuntimeError("Service is already running.")
 
         self.status = ProcessStatus.WORKING
+
         try:
-            if retry:
-                remaining_images: List[ImageMetadata] = [
-                    img
-                    for img in self.images
-                    if img.full_client_path not in self.processed_images_names
-                ]
-            else:
-                remaining_images: List[ImageMetadata] = [
-                    img
-                    for img in self.images
-                    if img.full_client_path not in self.processed_images_names
-                    and img.full_client_path not in self.failed_images_names
-                ]
-            index = 0
-            while index < len(remaining_images):
+            while True:
                 if self._terminate:
                     self.persist_progress()
                     break
-                start_time = time.perf_counter()
-                image = remaining_images[index]
 
-                for i in range(1):
-                    try:
-                        face_vectors = await self.process_image(image.full_client_path)
-                        break
-                    except RuntimeError as err:
-                        logger.warning(
-                            f"Failed to process image: {image.full_client_path} Attempt {i+1}. {err}. Restarting the httpx client"
-                        )
-                        await self._httpx_client.aclose()
-                        self._httpx_client = httpx.AsyncClient()
-                else:  # If all retries fail, execute this block
-                    logger.error(
-                        f"Failed to process image {image.full_client_path} after {i+1} attempts. Skipping image"
-                    )
-                    self.failed_images_names.append(image.full_client_path)
-                    self.persist_progress()
-                    index += 1
-                    continue
+                # Fetch one image to process
+                image = self._get_remaining_image(retry)
+                if not image:
+                    # No remaining images to process
+                    break
+                logger.info(f"Process image {image.name}")
+                image_process_status = await self._process_image(image)
 
-                for face in face_vectors["insights"]:
-                    # Extract image details and embedding
-                    face_object = Face(
-                        image_full_path=image.full_client_path,
-                        bbox=[int(coordinate) for coordinate in face["bbox"]],
-                        embedding=face["embedding"],
-                        ron_in_image=image.ron_in_image,
-                    )
-                    self.redis_interface.add_embedding(face_object)
-                    self._face_db_service.add_face(face_object)
-
-                self.processed_images_names.append(image.full_client_path)
-                logger.info(
-                    f"Processed {image.full_client_path} and extracted {len(face_vectors['insights'])} faces"
+                # Update progress dynamically
+                self._update_progress(
+                    image_name=image.full_client_path, image_status=image_process_status
                 )
-                process_time = time.perf_counter() - start_time
-                self._update_process_time(process_time)
-                # Update progress
-                processed_count = len(self.processed_images_names)
-                self.progress = int(processed_count / len(self.images) * 100)
+                self.update_image_status(
+                    full_client_path=image.full_client_path,
+                    new_status=image_process_status,
+                )
                 self.persist_progress()
 
-                # Recompute remaining_images to reflect changes dynamically
-                remaining_images = [
-                    img for img in self.images if img not in self.processed_images_names
-                ]
-                index += 1
+            self.status = (
+                ProcessStatus.DONE
+                if not self._get_remaining_image(False)
+                else ProcessStatus.IDLE
+            )
 
-            if len(remaining_images) == 0:
-                self.status = ProcessStatus.DONE
-            else:
-                self.status = ProcessStatus.IDLE
         except Exception as e:
             self.status = ProcessStatus.CRASHED
             self.persist_progress()
-            print(f"Error: {e}")
+            logger.error(f"Error occurred during processing: {e}")
+
         finally:
+            self._db.storage.flush()
             self._processing_task = None
+
+    def _get_remaining_image(self, retry: bool) -> Optional[ImageMetadata]:
+        """
+        Returns one image that still needs to be processed.
+        """
+        # logger.info("Fetching one remaining image with appropriate status.")
+        query = Query()
+        if not retry:
+            # Fetch one image where status is PENDING
+            result = self._db.get(
+                query.face_recognition_status
+                == ImageFaceRecognitionStatus.PENDING.value
+            )
+        else:
+            # Fetch one image where status is PENDING or FAILED
+            result = self._db.get(
+                (
+                    query.face_recognition_status
+                    == ImageFaceRecognitionStatus.PENDING.value
+                )
+                | (
+                    query.face_recognition_status
+                    == ImageFaceRecognitionStatus.FAILED.value
+                )
+            )
+
+        # Return the result as an ImageMetadata object or None if no match
+        return ImageMetadata(**result) if result else None
+
+    def _get_remaining_images(self, retry: bool) -> List[ImageMetadata]:
+        """
+        Returns the list of images that still need to be processed.
+        """
+        logger.info("Fetching remaining images with appropriate status.")
+        query = Query()
+        if not retry:
+            # Fetch images where status is PENDING
+            result = self._db.search(
+                query.face_recognition_status
+                == ImageFaceRecognitionStatus.PENDING.value
+            )
+        else:
+            # Fetch images where status is PENDING or FAILED
+            result = self._db.search(
+                (query.face_recognition_status != ImageFaceRecognitionStatus.DONE.value)
+                | (
+                    query.face_recognition_status
+                    == ImageFaceRecognitionStatus.FAILED.value
+                )
+            )
+
+        # Convert the results back to ImageMetadata objects
+        return [ImageMetadata(**doc) for doc in result]
+
+    async def _process_image(self, image: ImageMetadata) -> ImageFaceRecognitionStatus:
+        """
+        Processes a single image and handles retries.
+        """
+        start_time = time.perf_counter()
+
+        try:
+            face_vectors = await self._retry_process_image(image)
+            self._store_face_vectors(image, face_vectors)
+            self.processed_images_names.append(image.full_client_path)
+            logger.info(f"Processed {image.full_client_path} successfully.")
+            return ImageFaceRecognitionStatus.DONE
+        except RuntimeError as err:
+            self.failed_images_names.append(image.full_client_path)
+            logger.error(f"Failed to process image {image.full_client_path}: {err}")
+            return ImageFaceRecognitionStatus.FAILED
+        finally:
+            process_time = time.perf_counter() - start_time
+            self._update_process_time(process_time)
+
+    def _update_progress(
+        self, image_name: str, image_status: ImageFaceRecognitionStatus
+    ):
+        """
+        Updates the overall progress percentage.
+        """
+        if image_status == ImageFaceRecognitionStatus.FAILED:
+            self.failed_images_names.append(image_name)
+        if image_status == ImageFaceRecognitionStatus.DONE:
+            self.processed_images_names.append(image_name)
+
+        processed_count = len(self.processed_images_names)
+        self.progress = int(processed_count / self.get_images_count() * 100)
+
+    def _store_face_vectors(self, image: ImageMetadata, face_vectors: dict):
+        """
+        Stores face vectors in Redis and the database.
+        """
+        for face in face_vectors["insights"]:
+            face_object = Face(
+                image_full_path=image.full_client_path,
+                bbox=[int(coordinate) for coordinate in face["bbox"]],
+                embedding=face["embedding"],
+                ron_in_image=image.ron_in_image,
+            )
+            self.redis_interface.add_embedding(face_object)
+            self._face_db_service.add_face(face_object)
 
     def _update_process_time(self, t: float):
         if len(self._process_time) > 10:
             self._process_time.pop(0)
         self._process_time.append(t)
+
+    async def _retry_process_image(
+        self, image: ImageMetadata, retries: int = 1
+    ) -> dict:
+        """
+        Retries processing an image for a specified number of attempts.
+        """
+        for attempt in range(1, retries + 1):
+            try:
+                return await self.process_image(image.full_client_path)
+            except RuntimeError as err:
+                logger.warning(
+                    f"Attempt {attempt} failed for {image.full_client_path}: {err}. Restarting HTTP client."
+                )
+                await self._httpx_client.aclose()
+                self._httpx_client = httpx.AsyncClient()
+        raise RuntimeError(
+            f"Failed to process {image.full_client_path} after {retries} retries."
+        )
 
     async def process_image(self, image: str):
         """
@@ -304,15 +431,51 @@ class FaceRecognitionService:
 
         return result
 
+    def update_image_status(
+        self, full_client_path: str, new_status: ImageFaceRecognitionStatus
+    ):
+        """
+        Updates the status of an image in TinyDB based on its full client path.
+
+        Args:
+            full_client_path (str): The unique path of the image to update.
+            new_status (ImageFaceRecognitionStatus): The new status to set for the image.
+
+        Returns:
+            bool: True if the update was successful, False otherwise.
+        """
+        logger.info(f"Updating status for image: {full_client_path} to {new_status}.")
+
+        query = Query()
+        updated_count = self._db.update(
+            {"face_recognition_status": new_status.value},
+            query.full_client_path == full_client_path,
+        )
+
+        if updated_count[0] > 0:
+            logger.info(f"Successfully updated status for image: {full_client_path}.")
+            return True
+        else:
+            logger.warning(
+                f"Failed to update status for image: {full_client_path}. Image not found."
+            )
+            return False
+
     def persist_progress(self):
         """
-        Persists the current progress to a pickle file.
+        Persists the current progress to TinyDB incrementally.
         """
-        progress_data = {
-            "images": [image.model_dump() for image in self.images],
-            "processed_images": self.processed_images_names,
-            "progress": self.progress,
-            "failed_images": self.failed_images_names,
-        }
-        with open(self._progress_file, "wb") as file:
-            pickle.dump(progress_data, file)
+        # logger.info("Persisting progress to TinyDB.")
+
+        # Save metadata as a unique document
+        progress_query = Query()
+        self._db.upsert(
+            {
+                "id": "progress_metadata",  # Unique ID for progress metadata
+                "processed_images": self.processed_images_names,
+                "progress": self.progress,
+                "failed_images": self.failed_images_names,
+            },
+            progress_query.id == "progress_metadata",
+        )
+        # # Save each image as a unique document